Combining clicks and and charts
